Use a singleton queue object for Jython communication
This means we can do a true Python "import" of the scan syntax
definitions. Without a singleton queue object, we had to manually place
a queue in the Python namespace from Java, which was a bit hacky.

Having said that, I'm not sure that this singleton approach is ideal
either.

http://jira.diamond.ac.uk/browse/DAQ-2