Prevent warning messages about missing methods
Service unset methods need to take an instance of the service as an
argument, otherwise they are not identified and the OSGi framework
logs warnings about the methods not existing.

Also, this commit prevents exceptions and error messages on application
shutdown which occurred if the AtomQueueService had not been started.