Fix bug where a scannable couldn't be added to an empty Monitors view
This was due to ScannableContentProvider.insert() silently doing nothing
if the current list of scannables was empty. Existing tests always had
some scannables, even if they weren't being shown, so didn't trigger the
bug.

Note that the test for this had to go into a new class, as it caused
side effects which caused subsequent tests to fail.